+++ lib/PhasarLLVM/ControlFlow/EntryFunctionUtils.cpp
@@ -1,4 +1,4 @@
-#include "phasar/PhasarLLVM/Utils/EntryFunctionUtils.h"
+#include "phasar/PhasarLLVM/ControlFlow/EntryFunctionUtils.h"
 
 #include "phasar/PhasarLLVM/DB/LLVMProjectIRDB.h"
 #include "phasar/Utils/Logger.h"
